requirements.txt
Django==2.0.7
pytz==2018.5
djangorestframework==3.8.2
redis==2.10.6
psycopg2-binary==2.7.5
djangorestframework-jwt==1.11.0
drf-yasg[validation]
Pillow
django-dynamic-preferences
rest_condition
